+++ ultros-db/src/entity/owned_ffxiv_character.rs
@@ -8,7 +8,7 @@ use serde::{Deserialize, Serialize};
 pub struct Model {
     #[sea_orm(primary_key, auto_increment = false)]
     pub ffxiv_character_id: i32,
-    pub discord_user_id: i32,
+    pub discord_user_id: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
